+++ stellar-dotnet-sdk/responses/TransactionResponse.cs
@@ -12,7 +12,7 @@ namespace stellar_dotnet_sdk.responses
         [JsonProperty(PropertyName = "hash")] public string Hash { get; private set; }
 
         [JsonProperty(PropertyName = "ledger")]
-        public long Ledger { get; private set; }
+        public int Ledger { get; private set; }
 
         [JsonProperty(PropertyName = "created_at")]
         public string CreatedAt { get; private set; }
